+++ src/components/ResetPassword/__test__/emailSent.test.js
