+++ src/components/ResetPassword/__test__/skeleton.test.js
